A simple start on the updateindex command
Trying to reuse methods from invalidUsernames command
