+++ bin/lib/cli.js
@@ -3,6 +3,7 @@ const loadInit = require('./init')
 const loadStart = require('./start')
 const loadInvalidUsernames = require('./invalidUsernames')
 const loadMigrateLegacyResources = require('./migrateLegacyResources')
+const loadUpdateIndex = require('./updateIndex')
 const { spawnSync } = require('child_process')
 const path = require('path')
 
@@ -13,6 +14,7 @@ module.exports = function startCli (server) {
   loadStart(program, server)
   loadInvalidUsernames(program)
   loadMigrateLegacyResources(program)
+  loadUpdateIndex(program)
 
   program.parse(process.argv)
   if (program.args.length === 0) program.help()
